requirements.txt
discord.py
psutil
timeago==1.0.13
requests
bs4
html5lib
dislash.py